tetragon: Move data object to observer
Moving data object to observer. It helps following change
not to create golang dependency cycles when initializing
data cache.
